Svetetech/requirements.txt
django == 4.2.2
djangorestframework == 3.14.0
psycopg2-binary == 2.9.6
ipython == 8.12.2
djangorestframework-jwt
